# build/wiring-classes/META-INF/resources/quarkus-generated-openapi-doc.JSON
{
  "openapi" : "3.0.1",
  "info" : {
    "title" : "Generated API",
    "version" : "1.0"
  },
  "paths" : {
...
    "/sprints" : {
      "get" : {
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "application/json" : { }
            }
          }
        }
      },
      "post" : {
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/SprintRM"
              }
            }
          }
        },
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "application/json" : { }
            }
          }
        }
      }
    },
...
    "/sprints/{playerID}/{sprintID}" : {
      "get" : {
        "parameters" : [ {
          "name" : "playerID",
          "in" : "path",
          "required" : true,
          "schema" : {
            "$ref" : "#/components/schemas/Long"
          }
        }, {
          "name" : "sprintID",
          "in" : "path",
          "required" : true,
          "schema" : {
            "$ref" : "#/components/schemas/Long"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "application/json" : { }
            }
          }
        }
      }
    }
  },
  "components" : {
    "schemas" : {
...
      "Long" : {
        "format" : "int64",
        "type" : "integer"
      },
      "SprintRM" : {
        "type" : "object",
        "properties" : {
          "endDate" : {
            "format" : "date",
            "type" : "string"
          },
          "name" : {
            "type" : "string"
          },
          "sprintNumber" : {
            "format" : "int64",
            "type" : "integer"
          },
          "startDate" : {
            "format" : "date",
            "type" : "string"
          }
        }
      }
    }
  }
}